move logic out of split into epics - for undo funcitionality also add guards if none present and remove an undo if successful from the register
